fixed and refactored components
