+++ src/components/Header/Header.styles.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const StyledHeader = styled.header`
+export const StyledHeader = styled.nav`
 	padding: 20px 40px;
 	width: 100%;
 	display: flex;
