--- src/stories/CountriesSearchAndSelectModule.stories.tsx
@@ -1,8 +1,0 @@
-import { CountriesSearchAndSelectModule } from "../components/CountriesSearchAndSelectModule/CountriesSearchAndSelectModule";
-
-export default {
-	title: "CountriesSearchAndSelectModule",
-	component: CountriesSearchAndSelectModule,
-};
-
-export const Default = () => <CountriesSearchAndSelectModule />;
