+++ src/stories/FormPanel.stories.tsx
@@ -1,0 +1,8 @@
+import { FormPanel } from "../components/FormPanel/FormPanel";
+
+export default {
+	title: "FormPanel",
+	component: FormPanel,
+};
+
+export const Default = () => <FormPanel />;
